Move common azure properties into own file
