+++ packages/blocks/azure/src/lib/actions/azure-cli-action.ts
@@ -1,12 +1,8 @@
 import { createAction, Property } from '@openops/blocks-framework';
-import {
-  azureAuth,
-  dryRunCheckBox,
-  getAzureSubscriptionsStaticDropdown,
-} from '@openops/common';
-import { logger, SharedSystemProp, system } from '@openops/server-shared';
-import { getAzureErrorMessage } from '../error-helper';
-import { runCommand } from './azure-cli';
+import { azureAuth, dryRunCheckBox } from '@openops/common';
+import { logger } from '@openops/server-shared';
+import { runCommand } from '../azure-cli';
+import { subDropdown, useHostSession } from '../common-properties';
 
 export const azureCliAction = createAction({
   auth: azureAuth,
@@ -14,82 +10,8 @@ export const azureCliAction = createAction({
   description: 'Execute Azure CLI command',
   displayName: 'Azure CLI',
   props: {
-    useHostSession: Property.DynamicProperties({
-      displayName: '',
-      required: true,
-      refreshers: [],
-      props: async () => {
-        const enableHostSession = system.getBoolean(
-          SharedSystemProp.ENABLE_HOST_SESSION,
-        );
-
-        if (!enableHostSession) {
-          return {};
-        }
-
-        const result: any = {
-          useHostSessionCheckbox: Property.Checkbox({
-            displayName: 'Use host machine Azure CLI session',
-            description: `(Advanced) Uses the host machine's Azure CLI session. Requires 'az login' to have been run on the machine.`,
-            required: false,
-          }),
-        };
-
-        return result;
-      },
-    }),
-    subscriptions: Property.DynamicProperties({
-      displayName: '',
-      required: true,
-      refreshers: [
-        'auth',
-        'useHostSession',
-        'useHostSession.useHostSessionCheckbox',
-      ],
-      props: async ({ auth, useHostSession }) => {
-        let subDropdown;
-        try {
-          if (
-            useHostSession?.['useHostSessionCheckbox'] as unknown as boolean
-          ) {
-            subDropdown = await getSubscriptionsDropdownForHostSession(auth);
-          } else {
-            if (!auth) {
-              return {
-                subDropdown: Property.StaticDropdown({
-                  displayName: 'Subscriptions',
-                  description: 'Select a single subscription from the list',
-                  required: true,
-                  options: {
-                    disabled: true,
-                    options: [],
-                    placeholder: 'Please authenticate first',
-                  },
-                }),
-              };
-            }
-
-            subDropdown = await getAzureSubscriptionsStaticDropdown(auth);
-          }
-        } catch (error) {
-          subDropdown = Property.StaticDropdown({
-            displayName: 'Subscriptions',
-            description: 'Select a single subscription from the list',
-            required: true,
-            options: {
-              disabled: true,
-              options: [],
-              placeholder: `Something went wrong fetching subscriptions`,
-              error: getAzureErrorMessage(error),
-            },
-          });
-        }
-
-        return {
-          subDropdown: subDropdown,
-        };
-      },
-    }),
+    useHostSession: useHostSession,
+    subscriptions: subDropdown,
     commandToRun: Property.LongText({ displayName: 'Command', required: true }),
     dryRun: dryRunCheckBox(),
   },
@@ -128,27 +50,3 @@ export const azureCliAction = createAction({
     }
   },
 });
-
-async function getSubscriptionsDropdownForHostSession(auth: any) {
-  const result = await runCommand(
-    'account list --only-show-errors',
-    auth,
-    true,
-    undefined,
-  );
-
-  const parsedSubscriptions = JSON.parse(result);
-
-  return Property.StaticDropdown({
-    displayName: 'Subscriptions',
-    description: 'Select a single subscription from the list',
-    required: true,
-    options: {
-      disabled: false,
-      options: parsedSubscriptions.map((obj: { id: string; name: string }) => ({
-        label: obj.name,
-        value: obj.id,
-      })),
-    },
-  });
-}
