+++ packages/blocks/azure/src/lib/azure-cli.ts
